seeder user+bookmark

+++ src/modules/bookmarks/bookmarks.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { PageResponseDto } from '../pagination/dto/page-response.dto';
@@ -6,12 +6,16 @@ import { CreateBookmarkDto } from './dto/create-bookmark.dto';
 import { PageMetaDto } from '../pagination/dto/page-meta.dto';
 import { PageService } from '../pagination/page.service';
 import { Bookmark } from 'src/entities/bookmark.entity';
+import { Friend } from 'src/entities/friend.entity';
+import { FriendStatusValue } from 'src/commons/enums/friend/status-enum';
 
 @Injectable()
 export class BookmarksService extends PageService {
   constructor(
     @InjectRepository(Bookmark)
     private bookmarksRepository: Repository<Bookmark>,
+    @InjectRepository(Friend)
+    private friendRepository: Repository<Friend>,
   ) {
     super();
   }
@@ -21,6 +25,10 @@ export class BookmarksService extends PageService {
   async create(createBookmarkDto: CreateBookmarkDto, userId: number): Promise<PageResponseDto<Bookmark>> {
     const { ...params } = createBookmarkDto;
     const findBookmarks = await this.findBookmarks(userId, params.receiver_id);
+    const friend = await this.friendRepository.findOneBy({ sender_id: userId, receiver_id: params.receiver_id });
+    if (!friend || friend.status === FriendStatusValue.DA_DONG_Y.toString()) {
+      throw new BadRequestException('Các bạn chưa là bạn bè, không thể bookmark');
+    }
     if (findBookmarks) {
       await this.bookmarksRepository.delete(findBookmarks.id);
       return new PageResponseDto(findBookmarks);
